requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fsspec==2023.4.0
hydra-core==1.3.2
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.9
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.1.2+cu118
torch-summary==1.4.5
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.1.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.13
